Maven files structure modify and integration test added with Selenium
'mvn clean integration-test' know compile, run unit test, create war file, start Tomcat server, deplay war file on th Tomcat server, start Selenium server and run integration test.

+++ src/test/java/com/excilys/beans/TestComputer.java
@@ -12,30 +12,30 @@ import org.junit.Test;
 
 public class TestComputer {
 	private Computer computer;
-	
+
 	@After
 	public void tearDown() throws Exception {
 		computer = null;
 	}
-	
+
 	@Test
 	@Ignore("Dead code warning")
 	public void testEmpty() {
 		computer = new Computer();
-		
+
 		assertNull(computer.getId());
 		assertNull(computer.getName());
 		assertNull(computer.getDiscontinued());
 		assertNull(computer.getId());
 		assertNull(computer.getCompanyId());
 	}
-	
+
 	@Test
 	public void testNotEmpty1() {
 		String name = "test1";
 		Timestamp timestamp1 = new Timestamp(14545454);
 		Long companyId = 4L;
-		
+
 		computer = new Computer(name, timestamp1.toLocalDateTime().toLocalDate(), null, companyId);
 		assertNotNull(computer);
 		assertEquals(name, computer.getName());
@@ -43,17 +43,17 @@ public class TestComputer {
 		assertNull(computer.getDiscontinued());
 		assertEquals(companyId, computer.getCompanyId());
 	}
-	
+
 	@Test
 	public void testNotEmpty2() {
 		Long id = 3L;
 		String name = "test1";
 		Timestamp timestamp1 = new Timestamp(14545454);
 		Long companyId = 4L;
-		
+
 		computer = new Computer(id, name, null, timestamp1.toLocalDateTime().toLocalDate(), companyId);
 		assertNotNull(computer);
-		assertEquals(id.longValue(), computer.getId());
+		assertEquals(id, computer.getId());
 		assertEquals(name, computer.getName());
 		assertNull(computer.getIntroduced());
 		assertEquals(timestamp1.toLocalDateTime().toLocalDate(), computer.getDiscontinued());
